Crate bmw_ser

The BMW Serialization crate

This crate includes the crate::Serializable trait, the crate::Reader trait and the crate::Writer trait. They are separated from the bmw_util crate so that the util crate does not have to be a dependency of bmw_derive and can therefore use the Serializable proc_macro that is included in that crate. The Serializable trait is the key to several of the data structures in the bmw_util crate. It allows a specific way for data to be serialized so that it can be stored in various forms. The Reader and Writer traits are abstractions for reading and writing serializable data structures. The crate::Serializable macro is implemented for several data structures in this crate as well.

Structs

BinReader

Utility wrapper for an underlying byte Reader. Defines higher level methods to write numbers, byte vectors, hashes, etc.

BinWriter

Utility wrapper for an underlying byte Writer. Defines higher level methods to write numbers, byte vectors, hashes, etc.

Traits

Reader

Reader trait used for deserializing data.

Serializable

This is the trait used by all data structures to serialize and deserialize data. Anything stored in them must implement this trait. Commonly needed implementations are built in the ser module in this crate. These include Vec, String, integer types among other things.

Writer

Writer trait used for serializing data.

Functions

deserialize

Deserializes a Serializable from any std::io::Read implementation.

serialize

Serializes a Serializable into any std::io::Write implementation.