Trait bmw_ser::Writer

pub trait Writer {

    // Required method
    fn write_fixed_bytes<T: AsRef<[u8]>>(
        &mut self,
        bytes: T,
    ) -> Result<(), Error>;

    // Provided methods
    fn write_u8(&mut self, n: u8) -> Result<(), Error> { ... }
    fn write_i8(&mut self, n: i8) -> Result<(), Error> { ... }
    fn write_u16(&mut self, n: u16) -> Result<(), Error> { ... }
    fn write_i16(&mut self, n: i16) -> Result<(), Error> { ... }
    fn write_u32(&mut self, n: u32) -> Result<(), Error> { ... }
    fn write_i32(&mut self, n: i32) -> Result<(), Error> { ... }
    fn write_u64(&mut self, n: u64) -> Result<(), Error> { ... }
    fn write_i128(&mut self, n: i128) -> Result<(), Error> { ... }
    fn write_u128(&mut self, n: u128) -> Result<(), Error> { ... }
    fn write_i64(&mut self, n: i64) -> Result<(), Error> { ... }
    fn write_usize(&mut self, n: usize) -> Result<(), Error> { ... }
    fn write_bytes<T: AsRef<[u8]>>(&mut self, bytes: T) -> Result<(), Error> { ... }
    fn write_empty_bytes(&mut self, length: usize) -> Result<(), Error> { ... }

}

Writer trait used for serializing data.

Required Methods

fn write_fixed_bytes<T: AsRef<[u8]>>(&mut self, bytes: T) -> Result<(), Error>

write bytes to the stream

Provided Methods

fn write_u8(&mut self, n: u8) -> Result<(), Error>

write a u8 to the stream

fn write_i8(&mut self, n: i8) -> Result<(), Error>

write an i8 to the stream

fn write_u16(&mut self, n: u16) -> Result<(), Error>

write a u16 to the stream

fn write_i16(&mut self, n: i16) -> Result<(), Error>

write an i16 to the stream

fn write_u32(&mut self, n: u32) -> Result<(), Error>

write a u32 to the stream

fn write_i32(&mut self, n: i32) -> Result<(), Error>

write an i32 to the stream

fn write_u64(&mut self, n: u64) -> Result<(), Error>

write a u64 to the stream

fn write_i128(&mut self, n: i128) -> Result<(), Error>

write an i128 to the stream

fn write_u128(&mut self, n: u128) -> Result<(), Error>

write a u128 to the stream

fn write_i64(&mut self, n: i64) -> Result<(), Error>

write an i64 to the stream

fn write_usize(&mut self, n: usize) -> Result<(), Error>

write a usize to the stream

fn write_bytes<T: AsRef<[u8]>>(&mut self, bytes: T) -> Result<(), Error>

write bytes to the stream and specify the length so that variable length data may be written to the stream

fn write_empty_bytes(&mut self, length: usize) -> Result<(), Error>

write length empty (0u8) bytes to a stream.

Object Safety

This trait is not object safe.