Trait bmw_ser::Serializable

pub trait Serializable {
    fn read<R: Reader>(reader: &mut R) -> Result<Self, Error>
    where
        Self: Sized;
    fn write<W: Writer>(&self, writer: &mut W) -> Result<(), Error>;
}

This is the trait used by all data structures to serialize and deserialize data. Anything stored in them must implement this trait. Commonly needed implementations are built in the ser module in this crate. These include Vec, String, integer types among other things.

Required Methods

fn read<R: Reader>(reader: &mut R) -> Result<Self, Error> where
Self: Sized,

read data from the reader and build the underlying type represented by that data.

fn write<W: Writer>(&self, writer: &mut W) -> Result<(), Error>

write data to the writer representing the underlying type.